// clients/contentService/models/BlogPostCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BlogPostCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    tenantId?: string | null;
    enrolmentId?: string | null;
    title?: string | null;
    authorId?: string | null;
    description?: string | null;
    htmlContent?: string | null;
    featuredImageUrl?: string | null;
    blogPostCategoryID?: string | null;
    webTemplateID?: string | null;
};